+++ posthog-core/src/types.ts
@@ -49,6 +49,8 @@ export enum PostHogPersistedProperty {
   Props = 'props',
   FeatureFlags = 'feature_flags',
   FeatureFlagPayloads = 'feature_flag_payloads',
+  BootstrapFeatureFlags = 'bootstrap_feature_flags',
+  BootstrapFeatureFlagPayloads = 'bootstrap_feature_flag_payloads',
   OverrideFeatureFlags = 'override_feature_flags',
   Queue = 'queue',
   OptedOut = 'opted_out',
@@ -58,6 +60,8 @@ export enum PostHogPersistedProperty {
   GroupProperties = 'group_properties',
   InstalledAppBuild = 'installed_app_build', // only used by posthog-react-native
   InstalledAppVersion = 'installed_app_version', // only used by posthog-react-native
+  SessionReplay = 'session_replay', // only used by posthog-react-native
+  DecideEndpointWasHit = 'decide_endpoint_was_hit', // only used by posthog-react-native
 }
 
 export type PostHogFetchOptions = {
@@ -116,7 +120,11 @@ export type PostHogDecideResponse = {
     [key: string]: JsonType
   }
   errorsWhileComputingFlags: boolean
-  sessionRecording: boolean
+  sessionRecording?:
+    | boolean
+    | {
+        [key: string]: JsonType
+      }
 }
 
 export type PostHogFlagsAndPayloadsResponse = {