+++ CHANGELOG.md
@@ -5,3 +5,5 @@ Find the changelogs for the respective libraries in their folders:
 1. [posthog-node](https://github.com/PostHog/posthog-js-lite/blob/main/posthog-node/CHANGELOG.md)
 2. [posthog-react-native](https://github.com/PostHog/posthog-js-lite/blob/main/posthog-react-native/CHANGELOG.md)
 3. [posthog-web](https://github.com/PostHog/posthog-js-lite/blob/main/posthog-web/CHANGELOG.md)
+
+For Session Replay, See [posthog-react-native-session-replay](https://github.com/PostHog/posthog-react-native-session-replay/CHANGELOG.md)